validate providerid argument before attempting delete (aws, digitalocean, joyent, openstack, rackspace)

+++ lib/provisioner/worker/plugins/providers/fog_provider/lib/fog_provider/aws.rb
@@ -268,8 +268,12 @@ class FogProviderAWS < Provider
       # Delete server
       log.debug 'Invoking server delete'
       begin
+        raise ArgumentError if providerid.nil? || providerid.empty?
         server = connection.servers.get(providerid)
         server.destroy
+      rescue ArgumentError
+        log.debug "Invalid provider id #{providerid} specified on delete... skipping"
+      end
       rescue NoMethodError
         log.warn "Could not locate server '#{providerid}'... skipping"
       end
