+++ lib/provisioner/worker/plugins/providers/fog_provider/lib/fog_provider/digitalocean.rb
@@ -180,8 +180,11 @@ class FogProviderDigitalOcean < Provider
       # Delete server
       log.debug 'Invoking server delete'
       begin
+        raise ArgumentError if providerid.nil? || providerid.empty?
         server = connection.servers.get(providerid)
         server.destroy
+      rescue ArgumentError
+        log.debug "Invalid provider id #{providerid} specified on delete... skipping"
       rescue NoMethodError
         log.warn "Could not locate server '#{providerid}'... skipping"
       end
