+++ lib/provisioner/worker/plugins/providers/fog_provider/lib/fog_provider/joyent.rb
@@ -225,8 +225,11 @@ class FogProviderJoyent < Provider
       # Delete server
       log.debug 'Invoking server delete'
       begin
+        raise ArgumentError if providerid.nil? || providerid.empty?
         server = connection.servers.get(providerid)
         server.destroy
+      rescue ArgumentError
+        log.debug "Invalid provider id #{providerid} specified on delete... skipping"
       rescue NoMethodError, Fog::Compute::Joyent::Errors::NotFound
         log.warn "Could not locate server '#{providerid}'... skipping"
       else
